[Refactor] : 기존 data 파일 클린아키텍처 패턴으로 리팩토링 (#54)
