+++ app/src/main/java/com/example/cherry_pick_android/data/data/Keyword.kt
@@ -1,0 +1,5 @@
+package com.example.cherry_pick_android.data.data
+
+data class Keyword(
+    var keyword: String
+)
