+++ app/src/main/java/com/example/cherry_pick_android/data/data/News.kt
@@ -1,4 +1,5 @@
 package com.example.cherry_pick_android.data.data
 
-class Exam {
-}
+data class News(
+    val news: String
+)
