+++ app/src/main/java/com/example/cherry_pick_android/data/mapper/KeywordMapper.kt
@@ -1,8 +1,6 @@
 package com.example.cherry_pick_android.data.mapper
 
-import androidx.lifecycle.LiveData
 import com.example.cherry_pick_android.data.model.KeywordEntity
-import com.example.cherry_pick_android.domain.model.Keyword
 
 // 키워드 매핑 (domain 계층에 존재하는 키워드 데이터를 data 계층에서 사용하기 위한 것)
 object KeywordMapper {
