+++ app/src/main/java/com/example/cherry_pick_android/data/remote/request/KeywordDAO.kt
@@ -1,4 +1,4 @@
-package com.example.cherry_pick_android.data.db
+package com.example.cherry_pick_android.data.remote.request
 
 import androidx.lifecycle.LiveData
 import androidx.room.Dao
