+++ app/src/main/java/com/example/cherry_pick_android/data/remote/response/ResponseExamData.kt
@@ -1,0 +1,7 @@
+package com.example.cherry_pick_android.data.remote.response
+
+data class ResponseExamData(
+    val result: Void
+)//: BaseResponse(){
+//
+// }
