--- app/src/main/java/com/example/cherry_pick_android/data/remote/service/Exam.kt
@@ -1,4 +1,0 @@
-package com.example.cherry_pick_android.data.remote.service
-
-class Exam {
-}
