+++ app/src/main/java/com/example/cherry_pick_android/data/remote/service/ExamService.kt
@@ -1,0 +1,11 @@
+package com.example.cherry_pick_android.data.remote.service
+
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface ExamService {
+    // @GET("/example")
+    // suspend fun getExamData(
+    //    @Query("exam") exam: Int
+    // ) : Response<ResponseExamData>
+}
