+++ app/src/main/java/com/example/cherry_pick_android/data/remote/source/ExamRemoteSource.kt
@@ -1,4 +1,4 @@
 package com.example.cherry_pick_android.data.remote.source
 
-class Exam {
+class ExamRemoteSource {
 }
