--- app/src/main/java/com/example/cherry_pick_android/domain/model/Keyword.kt
@@ -1,5 +1,0 @@
-package com.example.cherry_pick_android.domain.model
-
-data class Keyword(
-    var keyword: String
-)
