--- app/src/main/java/com/example/cherry_pick_android/domain/model/ScrapNews.kt
@@ -1,5 +1,0 @@
-package com.example.cherry_pick_android.domain.model
-
-data class ScrapNews(
-    val scrapNews: String
-)
