--- app/src/main/java/com/example/cherry_pick_android/domain/repository/Exam.kt
@@ -1,2 +1,0 @@
-package com.example.cherry_pick_android.domain.repository
-
