+++ app/src/main/java/com/example/cherry_pick_android/domain/source/ExamSource.kt
@@ -1,4 +1,4 @@
 package com.example.cherry_pick_android.domain.source
 
-class Exam {
+interface ExamSource {
 }
