--- app/src/main/java/com/example/cherry_pick_android/domain/usecase/Exam.kt
@@ -1,4 +1,0 @@
-package com.example.cherry_pick_android.domain.usecase
-
-class Exam {
-}
