--- app/src/main/java/com/example/cherry_pick_android/presentation/adapter/ExampleViewHolder.kt
@@ -1,5 +1,0 @@
-package com.example.cherry_pick_android.presentation.adapter
-
-class ExampleViewHolder(
-) {
-}
