+++ app/src/main/java/com/example/cherry_pick_android/presentation/di/RepositoryModule.kt
@@ -1,8 +1,7 @@
-package com.example.cherry_pick_android.data.di
+package com.example.cherry_pick_android.presentation.di
 
-import com.example.cherry_pick_android.data.KeywordRepositoryImpl
-import com.example.cherry_pick_android.data.db.KeywordDAO
-import com.example.cherry_pick_android.data.db.KeywordDB
+import com.example.cherry_pick_android.data.remote.repository.KeywordRepositoryImpl
+import com.example.cherry_pick_android.data.remote.request.KeywordDAO
 import com.example.cherry_pick_android.domain.repository.KeywordRepository
 import dagger.Module
 import dagger.Provides
