+++ app/src/main/java/com/example/cherry_pick_android/presentation/ui/searchList/SearchListActivity.kt
@@ -6,7 +6,7 @@ import androidx.appcompat.app.AppCompatActivity
 import com.example.cherry_pick_android.databinding.ActivitySearchListBinding
 import com.example.cherry_pick_android.presentation.adapter.NewsRecyclerViewAdapter
 import com.example.cherry_pick_android.presentation.ui.article.ArticleActivity
-import com.example.cherry_pick_android.domain.model.News
+import com.example.cherry_pick_android.data.data.News
 
 class SearchListActivity : AppCompatActivity() {
     private lateinit var binding: ActivitySearchListBinding
